Move all views unders a single folder structure
