+++ src/Slider/Slider.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import './Slider.css';
-import View1 from '../View 1/View1';
-import View2 from '../View 2/View2';
-import View3 from '../View 3/View3';
-import View4 from '../View 4/View4';
-import View6 from '../View 6/View6';
+import View1 from '../Views/View 1/View1';
+import View2 from '../Views/View 2/View2';
+import View3 from '../Views/View 3/View3';
+import View4 from '../Views/View 4/View4';
+import View6 from '../Views/View 6/View6';
 import GoogleSlides from '../GoogleSlides/GoogleSlides';
 import config from '../config.json';
 import BlackoutPeriods from '../_components/blackout-periods/blackout-periods';
