+++ src/Views/View 1/View1.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
 import './View1.css';
-import Logo from '../_components/logo/logo';
-import Clock from '../_components/clock/clock';
-import PrayerTimes from '../_components/prayer-times/prayer-times';
-import Date from '../_components/date/date';
-import BuildNumber from '../_components/build-number/build-number';
-import SunriseAndZawwal from '../_components/sunrise-and-zawwal/sunrise-and-zawwal';
-import AdditionalMessage from '../_components/additional-message/additional-message';
-import AppConfig from '../_components/app-config/app-config';
-import Branding from '../_components/branding/branding';
+import Logo from '../../_components/logo/logo';
+import Clock from '../../_components/clock/clock';
+import PrayerTimes from '../../_components/prayer-times/prayer-times';
+import Date from '../../_components/date/date';
+import BuildNumber from '../../_components/build-number/build-number';
+import SunriseAndZawwal from '../../_components/sunrise-and-zawwal/sunrise-and-zawwal';
+import AdditionalMessage from '../../_components/additional-message/additional-message';
+import AppConfig from '../../_components/app-config/app-config';
+import Branding from '../../_components/branding/branding';
 
 class View1 extends Component {
   constructor(props) {
