+++ src/Views/View 4/View4.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import './View4.css';
-import Logo from '../_components/logo/logo';
-import Clock from '../_components/clock/clock';
-import BuildNumber from '../_components/build-number/build-number';
-import PrayerTimesWeekAhead from '../_components/prayer-times-week-ahead/prayer-times-week-ahead';
-import Branding from '../_components/branding/branding';
+import Logo from '../../_components/logo/logo';
+import Clock from '../../_components/clock/clock';
+import BuildNumber from '../../_components/build-number/build-number';
+import PrayerTimesWeekAhead from '../../_components/prayer-times-week-ahead/prayer-times-week-ahead';
+import Branding from '../../_components/branding/branding';
 
 class View4 extends Component {
   render() {
