+++ src/Views/View 5/View5.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import './View5.css';
-import Logo from '../_components/logo/logo';
-import Clock from '../_components/clock/clock';
-import BuildNumber from '../_components/build-number/build-number';
-import Branding from '../_components/branding/branding';
+import Logo from '../../_components/logo/logo';
+import Clock from '../../_components/clock/clock';
+import BuildNumber from '../../_components/build-number/build-number';
+import Branding from '../../_components/branding/branding';
 
 class View5 extends Component {
   componentWillMount() {
