+++ src/Views/View 6/View6.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import './View6.css';
-import Logo from '../_components/logo/logo';
-import Clock from '../_components/clock/clock';
-import BuildNumber from '../_components/build-number/build-number';
-import Branding from '../_components/branding/branding';
-import HadithOfTheDay from '../_components/hadith-of-the-day/hadith-of-the-day';
+import Logo from '../../_components/logo/logo';
+import Clock from '../../_components/clock/clock';
+import BuildNumber from '../../_components/build-number/build-number';
+import Branding from '../../_components/branding/branding';
+import HadithOfTheDay from '../../_components/hadith-of-the-day/hadith-of-the-day';
 
 class View6 extends Component {
   render() {
